refactor: controller and model refactoring rename class from CompareResult to ScoreBoard move method from controller(BaseBallGame) to model(ScoreBaord)
